+++ server/models/index.js
@@ -1,32 +1,35 @@
 
-
-const fs = require('fs');
-const path = require('path');
+const fs        = require('fs');
+const path      = require('path');
 const Sequelize = require('sequelize');
 
-const basename = path.basename(__filename);
-const env = process.env.NODE_ENV || 'development';
-const config = require(`${__dirname}/..\config.json`)[env];
-const db = {};
+const basename  = path.basename(module.filename);
+const env       = process.env.NODE_ENV || 'development';
+const config = require(`${__dirname}/../config/config.json`)[env];
+const db        = {};
 
+let sequelize;
 if (config.use_env_variable) {
-  var sequelize = new Sequelize(process.env[config.use_env_variable]);
+  sequelize = new Sequelize(process.env[config.use_env_variable]);
 } else {
-  var sequelize = new Sequelize(config.database, config.username, config.password, config);
+  sequelize = new Sequelize(config.database, config.username, config.password, config);
 }
 
 fs
-  .readdirSync(__dirname)
-  .filter(file => (file.indexOf('.') !== 0) && (file !== basename) && (file.slice(-3) === '.js'))
-  .forEach((file) => {
-    const model = sequelize.import(path.join(__dirname, file));
-    db[model.name] = model;
-  });
+.readdirSync(__dirname)
+.filter(file =>
+  (file.indexOf('.') !== 0) &&
+  (file !== basename) &&
+  (file.slice(-3) === '.js'))
+.forEach(file => {
+  const model = sequelize.import(path.join(__dirname, file));
+  db[model.name] = model;
+});
 
-Object.keys(db).forEach((modelName) => {
-  if (db[modelName].associate) {
-    db[modelName].associate(db);
-  }
+Object.keys(db).forEach(modelName => {
+if (db[modelName].associate) {
+  db[modelName].associate(db);
+}
 });
 
 db.sequelize = sequelize;
